Execute NOT NULL with ADD COLUMN in PG11+ (#39)
* Execute NOT NULL with ADD COLUMN in PG11+

One of the performance improvements in Postgres 11 was executing an `ADD COLUMN ... NOT NULL` without needing a table rewrite. This is now a safe operation in PG11+, and we can avoid adding backwards-compatible safe version with a statement timeout that can fail in some cases.

https://www.postgresql.org/docs/release/11.0/

* Fix variable naming

* Ignore Metrics/PerceivedComplexity for add_column

* Remove unnecessary check

* Add more tests around add_column for older PG

* Refactor add_column to early out for PG11+

This simplifies the if checks enough that we no longer have to ignore the Rubocop complaints about complexity.