Update logic for overriding :algorithm in add_index
Rails (at least version 7.1) raises an error if the `:algorithm` option is set to :default, before this code gets
invoked.

Instead of allowing a custom `:default` option, we can instead just check whether the options[:algorithm] key is
present. This is what the `remove_index` method below does, and allows us to use `algorithm: nil` to disable the
`:concurrently` option.